treewide: move rmsgdigest under crypto/
rmsgdigest is the cryptographic-hash entry point - SHA-1 / SHA-2,
soon-to-be SHAKE / SHA-3 - so it belongs alongside the rest of the
crypto headers rather than at the top level. Move the public header
to include/rlib/crypto/rmsgdigest.h and the implementation to
rlib/crypto/rmsgdigest.c; bump every #include and the meson source
list. rcrypto.h gets the include too so the crypto umbrella is
self-contained without depending on rlib.h.

rlib.h keeps the include so existing callers that pull in <rlib.h>
still get the digest API without changes.
