+++ include/rlib/net/proto/rsdp.h
@@ -26,7 +26,7 @@
 #include <rlib/rref.h>
 
 #include <rlib/rbuffer.h>
-#include <rlib/rmsgdigest.h>
+#include <rlib/crypto/rmsgdigest.h>
 #include <rlib/rsocketaddress.h>
 #include <rlib/rstr.h>
 #include <rlib/ruri.h>
