+++ include/rlib/rcrypto.h
@@ -107,6 +107,7 @@
 #include <rlib/crypto/recurve-montgomery.h>
 #include <rlib/crypto/rhmac.h>
 #include <rlib/crypto/rkey.h>
+#include <rlib/crypto/rmsgdigest.h>
 #include <rlib/crypto/rpem.h>
 #include <rlib/crypto/rrsa.h>
 #include <rlib/crypto/rxdh.h>
