+++ include/rlib/rlib.h
@@ -44,7 +44,7 @@
 #include <rlib/rmemallocator.h>
 #include <rlib/rmemfile.h>
 #include <rlib/rmodule.h>
-#include <rlib/rmsgdigest.h>
+#include <rlib/crypto/rmsgdigest.h>
 #include <rlib/rpoll.h>
 #include <rlib/rrand.h>
 #include <rlib/rref.h>
