+++ rlib/crypto/rpem.c
@@ -31,7 +31,7 @@
 #include <rlib/charset/rascii.h>
 
 #include <rlib/rmem.h>
-#include <rlib/rmsgdigest.h>
+#include <rlib/crypto/rmsgdigest.h>
 #include <rlib/rstr.h>
 #include <rlib/rbase64.h>
